% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datosfaltantes.R
\name{datos_faltantes}
\alias{datos_faltantes}
\title{Datos faltantes}
\usage{
datos_faltantes(dataframe1)
}
\arguments{
\item{dataframe1}{Dataframe exportado ó creado con datos externos}
}
\value{
mensaje de en que columna existen datos faltantes
}
\description{
Datos faltantes
}
\examples{
d <- data.frame(Nombre = c("Esteban", NA, "Beatriz"),
                Edad = c(2, 93, 57),
                Ciudad = factor(c(NA, "Granada", "Jaén")),
                Casado = c(FALSE, TRUE, NA))
datos_faltantes(d)

edad <- c(22, 34, 29, 25, NA)
tiempo <- c(14.21,NA, 11.89, 13.81, 12.03)
sexo <- c("M","H","H","M","M")
misDatos <- data.frame(edad,tiempo,sexo)
datos_faltantes(misDatos)

}
